+++ src/tests/mod.rs
@@ -8,3 +8,4 @@ pub mod test_schedule;
 pub mod test_scheduler;
 pub mod test_manager;
 pub mod test_command;
+pub mod test_agent;
