+++ src/worker.rs
@@ -38,48 +38,96 @@ pub struct WorkerTasks{
     handles: WorkerTasksHandles
 }
 
-
+/// Represents the status of a worker check.
+///
+/// This enum is used to determine whether a worker should continue monitoring 
+/// or stop its operations. It provides two possible states:
+/// - `Continue`: Indicates that the worker should continue monitoring.
+/// - `Stop`: Indicates that the worker should cease monitoring.
 #[derive(Debug,Clone, PartialEq)]
 pub enum WorkerCheckStatus {
     Continue,
     Stop
 }
 
-impl WorkerTasks{
-
+impl WorkerTasks {
+    /// Creates a new instance of `WorkerTasks`.
+    ///
+    /// Initializes the `WorkerTasks` struct with an empty collection of task handles.
+    ///
+    /// # Returns
+    /// A new `WorkerTasks` instance.
     pub async fn new() -> Self {
         Self {
-            handles: HashMap::new()
+            handles: HashMap::new(),
         }
     }
 
-    pub async fn to_id(id: tokio::task::Id) -> u64 {        
+    /// Converts a Tokio task ID into a `u64` value.
+    ///
+    /// # Arguments
+    /// * `id` - The Tokio task ID to be converted.
+    ///
+    /// # Returns
+    /// A `u64` representation of the given task ID.
+    pub async fn to_id(id: tokio::task::Id) -> u64 {
         id.to_string().parse::<u64>().unwrap()
     }
 
-    pub async fn to_name(queue_name: &String, runtime_id: &u64) -> String {        
-        format!("{}-{}",queue_name,runtime_id)
-    }
-    
-    pub async fn add(&mut self, queue_name: String, runtime_id: u64, handle: Option<JoinHandle<()>>) {
-        let name: String = Self::to_name(&queue_name, &runtime_id).await;
-        self.handles.insert(name.clone(),handle);
+    /// Generates a unique name for a task.
+    ///
+    /// # Arguments
+    /// * `queue_name` - A reference to the name of the queue the task belongs to.
+    /// * `task_id` - A unique identifier for the task.
+    ///
+    /// # Returns
+    /// A string combining the queue name and task ID in the format `{queue_name}-{task_id}`.
+    pub async fn to_name(queue_name: &String, task_id: &u64) -> String {
+        format!("{}-{}", queue_name, task_id)
     }
 
+    /// Adds a task to the worker's collection of handles.
+    ///
+    /// # Arguments
+    /// * `queue_name` - The name of the queue the task belongs to.
+    /// * `task_id` - A unique identifier for the task.
+    /// * `handle` - An optional join handle associated with the task.
+    ///
+    /// The task is stored in the collection under a unique name generated by combining
+    /// the queue name and task ID.
+    pub async fn add(&mut self, queue_name: String, task_id: u64, handle: Option<JoinHandle<()>>) {
+        let name: String = Self::to_name(&queue_name, &task_id).await;
+        self.handles.insert(name.clone(), handle);
+    }
+
+    /// Removes a task from the worker's collection by its unique name.
+    ///
+    /// # Arguments
+    /// * `name` - The unique name of the task to be removed.
     pub async fn remove(&mut self, name: String) {
         self.handles.remove(&name);
-    } 
-    
-    pub async fn abort(&self, name: String) -> Result<bool,String> {
+    }
+
+    /// Attempts to abort a running task by its unique name.
+    ///
+    /// # Arguments
+    /// * `name` - The unique name of the task to be aborted.
+    ///
+    /// # Returns
+    /// * `Ok(true)` if the task was successfully aborted.
+    /// * `Err(String)` with an error message if the task could not be found or aborted.
+    pub async fn abort(&self, name: String) -> Result<bool, String> {
         if let Some(handle) = self.handles.get(&name).clone() {
             if let Some(h) = handle {
                 h.abort();
-            } 
+            }
             return Ok(true);
         }
-        return Err(format!("unable to abort worker task {}",name));
+        Err(format!("unable to abort worker task {}", name))
     }
 }
+
+
 
 impl Worker {
     /// Creates a new instance of the `Worker`.
@@ -100,6 +148,22 @@ impl Worker {
             author
         }
     }
+
+    /// Checks for and executes commands related to task management.
+    ///
+    /// This function processes commands associated with worker tasks and queues, such as removing tasks, 
+    /// terminating tasks, and shutting down queues. It interacts with the provided runtime, worker tasks, 
+    /// and database components to perform the necessary operations.
+    ///
+    /// # Arguments
+    /// * `runtime` - A reference to the Tokio runtime, used to access runtime metrics.
+    /// * `worker_tasks` - A mutable reference to the `WorkerTasks` structure, which manages active task handles.
+    /// * `queue_name` - The name of the queue associated with the commands.
+    ///
+    /// # Returns
+    /// * `Ok(WorkerCheckStatus::Continue)` if the worker should continue monitoring the queue.
+    /// * `Ok(WorkerCheckStatus::Stop)` if the worker should stop monitoring the queue.
+    /// * `Err(String)` if an error occurs while processing commands.
     pub async fn check_command(&self, runtime: &Runtime, worker_tasks: &mut WorkerTasks, queue_name: String) -> Result<WorkerCheckStatus,String>{
         let queue: Queue = Queue::new(self.db.clone()).await;
         let agent: Agent = Agent::new(self.db.clone()).await;
@@ -211,7 +275,6 @@ impl Worker {
                 }).await?;
                 worker_tasks.add(queue_name.clone(), 0,  None).await;
                 loop {
-                    
                     match self.check_command(&runtime, &mut worker_tasks, queue_name.clone() ).await {
                         Ok(wc_status) => {
                             if wc_status == WorkerCheckStatus::Stop  {
@@ -309,7 +372,6 @@ impl Worker {
                                                                 }                                                         
                                                             }
                                                             Err(error) => {
-                                                                println!("Error: {}",error);
                                                                 if let Err(error) = queue.set_status(queue_id.clone(),QueueStatus::Error,Some(error.clone())).await {
                                                                     error!("task execution result error [{}]: {}", record_name, error);
                                                                 }
